+++ fpdf/html.py
@@ -736,9 +736,10 @@ class HTMLMixin:
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         warnings.warn(
-            # pylint: disable=implicit-str-concat
-            "The HTMLMixin class is deprecated since v2.6.0. "
-            "Simply use the FPDF class as a replacement.",
+            (
+                "The HTMLMixin class is deprecated since v2.6.0. "
+                "Simply use the FPDF class as a replacement."
+            ),
             DeprecationWarning,
             stacklevel=get_stack_level(),
         )