+++ test/encryption/test_encryption.py
@@ -176,7 +176,6 @@ def test_encrypt_font(tmp_path):
     )
     pdf.set_font("Quicksand", size=32)
     text = (
-        # pylint: disable=implicit-str-concat
         "Lorem ipsum dolor, **consectetur adipiscing** elit,"
         " eiusmod __tempor incididunt__ ut labore et dolore --magna aliqua--."
     )