+++ test/svg/test_svg.py
@@ -1,4 +1,4 @@
-# pylint: disable=implicit-str-concat, no-self-use, protected-access
+# pylint: disable=no-self-use, protected-access
 import io
 from pathlib import Path