+++ test/text/test_multi_cell_markdown.py
@@ -1,4 +1,3 @@
-# pylint: disable=implicit-str-concat
 from pathlib import Path
 
 import fpdf